Log env, move core galactus files, include latest utils
